+++ CHANGELOG.md
@@ -12,6 +12,8 @@ Versioning].
 
 - (`dbbd871`) Add option to disallow top-level property accesses to
   `no-top-level-side-effects`.
+- (`dbbd871`) Add option to allow adding properties to a function declaration at
+  the top-level `no-top-level-side-effects`.
 
 ## [3.4.1] - 2025-03-16